LayoutTests/http/tests/preload/download_resources-expected.txt

CONSOLE MESSAGE: line 13: <link rel=preload> must have a valid `as` value
This test makes sure that link preload preloads resources

PASS Makes sure that preloaded resources are downloaded 
